Changed short to int hashset because of rna mode

+++ src/main/java/blockteil/filters/SmallThresholdFilterer.java
@@ -16,10 +16,10 @@ public class SmallThresholdFilterer extends KMERFilterer {
         A: for (int i = 0; i < seq.length() - Config.KMER_LENGTH + 1; i += Config.OFFSET) {
             kmer = revcomp ? KMER.makeKMER_revcomp(seq, i) : KMER.makeKMER(seq, i);
 
-            Set<Short> matchingGenes = Config.KMER_MAP.get(kmer);
+            Set<Integer> matchingGenes = Config.KMER_MAP.get(kmer);
             if (matchingGenes == null) continue;
 
-            for (short gene : matchingGenes) {
+            for (int gene : matchingGenes) {
                 out.set(gene);
                 if (Config.EARLY_TERMINATION_ALLOWED) break A;
             }
